refactor: combine rpc-server and authorizer
