+++ index.js
@@ -1,5 +1,4 @@
 require('./src/rpc-server')
-require('./src/authorizer')
 require('./src/connector')
 require('./src/http-client')
 require('./src/quoter')
