+++ test/run.sh
@@ -1,6 +1,6 @@
-node ./src/rpc-server &
-node ./src/authorizer &
-node ./src/connector &
-node ./src/http-client &
-node ./src/quoter &
-node ./src/fulfiller &
+node ../src/rpc-server &
+node ../src/authorizer &
+node ../src/connector &
+node ../src/http-client &
+node ../src/quoter &
+node ../src/fulfiller &
